% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline.R
\name{read_cs_data}
\alias{read_cs_data}
\title{read_cs_data}
\usage{
read_cs_data(filename, return_option = "data")
}
\arguments{
\item{filename}{The path to the file containing the CSI TOA5 data.}

\item{return_option}{A string indicating what to return. Can be either "data" or
"info". If "data", the function will return a data table containing the
CSI TOA5 data. If "info", the function will return a list containing the
first four lines of the file, which contain metadata about the data.}
}
\value{
A data table containing the CSI TOA5 data, or a list containing the
   first four lines of the file, depending on the value of the `return_option`
   argument.
}
\description{
Read CSI TOA5 data from a file.

Reads CSI TOA5 data from a file and returns a data table.
}
\examples{
\dontrun{
read_cs_data("csi_toa5_data.csv")
# A data table containing the CSI TOA5 data

read_cs_data("csi_toa5_data.csv", return_option = "info")
# A list containing the first four lines of the file, which contain metadata about the data
}
}
