% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline-package.r
\docType{package}
\name{skyline-package}
\alias{skyline}
\alias{skyline-package}
\title{skyline: Analysis Of Automated Chamber Measurements Of GHG Fluxes}
\description{
This package contains code for the analysis of automated chamber measurements of GHG fluxes.
}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID}) [copyright holder]

}
\keyword{internal}
